+++ packages/moolah-lending-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @lista-dao/moolah-lending-sdk
 
+## 1.0.3
+
+### Patch Changes
+
+- Enhance Moolah SDK with holdings functionality and improve API parameter handling
+- Updated dependencies
+  - @lista-dao/moolah-sdk-core@1.0.4
+
 ## 1.0.2
 
 ### Patch Changes